Log In Page using html and CSS
Hi, This is a normal log in page designed using html and CSS. Though registration is mentioned but I want to clarify here that it has no functionality as javascript is not utilized. So, you can take the code as it is or take the part you want. Its up to you. Thanks.